user/pyproject.toml
[tool.poetry]
name = "user"
version = "0.1"
description = "user"
authors = ["LedxDeliveryFlopp <korstim18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
alembic = "^1.13.1"
email-validator = "^2.1.0.post1"
passlib = "^1.7.4"
pika = "^1.3.2"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
